Fix a possible method resolution bug

+++ v1_15_R1/src/main/java/net/citizensnpcs/nms/v1_15_R1/util/NMSImpl.java
@@ -615,7 +615,7 @@ public class NMSImpl implements NMSBridge {
                         }
                     }
                 }
-                navigation.a(params.speed());
+                navigation.a((double) params.speed());
                 return NMSImpl.isNavigationFinished(navigation);
             }
         };
